// features/api/model/auditLogIndex200.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Petrol Integrity System API Documentation
 * OpenAPI spec version: 0.0.1
 */
import type { AuditLogIndex200Links } from './auditLogIndex200Links';
import type { AuditLogIndex200Meta } from './auditLogIndex200Meta';
import type { AuditLogResource } from './auditLogResource';

export type AuditLogIndex200 = {
  data: AuditLogResource[];
  links: AuditLogIndex200Links;
  meta: AuditLogIndex200Meta;
};